fastapi/api/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-jose[cryptography]
httpx
pydantic-settings
python-dotenv
